invnetory app

+++ internal/item/item.go
@@ -59,9 +59,9 @@ func (c *Configuration) Amount(cat Category) int32 {
 			v, exists := c.Header["Amount"]
 			if exists {
 				core.AppLog.Printf("Amont %v:\n", v)
-				am, ok := v.(int32)
+				am, ok := v.(int64)
 				if ok {
-					return am
+					return int32(am)
 				}
 				return 0
 			}
